+++ core/agent_harness/llm_resolution.py
