--- core/agent_harness/providers/__init__.py
@@ -1,7 +1,0 @@
-"""Core-owned default port implementations and provider resolution.
-
-Default tool / prompt-context / reasoning providers usable by any surface, plus
-the LLM provider and model resolution shared across surfaces.
-"""
-
-from __future__ import annotations
