+++ core/agent_harness/turns/headless_dispatch.py
@@ -28,6 +28,7 @@ Example::
 
 from __future__ import annotations
 
+from core.agent_harness.accounting.turn_accounting import DefaultTurnAccounting
 from core.agent_harness.ports import (
     ConfirmFn,
     ErrorReporter,
@@ -39,11 +40,10 @@ from core.agent_harness.ports import (
     ToolProvider,
     TurnAccounting,
 )
-from core.agent_harness.providers.default_prompt_context import (
+from core.agent_harness.prompts.prompt_context import (
     DefaultPromptContextProvider,
     supports_default_prompt_context,
 )
-from core.agent_harness.providers.default_providers import DefaultTurnAccounting
 from core.agent_harness.turns.action_driver import run_action_agent_turn
 from core.agent_harness.turns.evidence_driver import gather_tool_evidence
 from core.agent_harness.turns.headless_adapters import (
