+++ gateway/tests/test_turn_handler.py
@@ -44,7 +44,7 @@ def test_turn_handler_resolves_action_tools_from_live_session(monkeypatch: Any) 
         return [MagicMock(name="slack_send_message")]
 
     monkeypatch.setattr(
-        "core.agent_harness.providers.default_providers.get_action_tools_from_integrations_context",
+        "core.agent_harness.tools.tool_provider.get_action_tools_from_integrations_context",
         _fake_get_tools,
     )
