+++ tests/core/agent/orchestration/test_model_question_misroute.py
@@ -23,8 +23,8 @@ from rich.console import Console
 
 import surfaces.interactive_shell.runtime.shell_turn_execution as shell_turn_execution
 import tools.interactive_shell.actions.slash as slash_tool
-from core.agent_harness.providers import default_prompt_context
-from core.agent_harness.providers.default_prompt_context import DefaultPromptContextProvider
+from core.agent_harness.prompts import prompt_context as default_prompt_context
+from core.agent_harness.prompts.prompt_context import DefaultPromptContextProvider
 from surfaces.interactive_shell.command_registry import dispatch_slash
 from surfaces.interactive_shell.session import Session
 from tests.core.agent.orchestration.action_execution_test_harness import (
