+++ tests/interactive_shell/chat/test_cli_agent.py
@@ -18,6 +18,7 @@ from typing import Any
 
 from rich.console import Console
 
+from core.agent_harness.prompts import prompt_context as default_prompt_context
 from core.agent_harness.prompts.assistant_agent_prompt import (
     _MARKDOWN_RULE,
     _TERMINOLOGY_RULE,
@@ -25,8 +26,7 @@ from core.agent_harness.prompts.assistant_agent_prompt import (
     _build_system_prompt,
     build_environment_block,
 )
-from core.agent_harness.providers import default_prompt_context
-from core.agent_harness.providers.default_prompt_context import DefaultPromptContextProvider
+from core.agent_harness.prompts.prompt_context import DefaultPromptContextProvider
 from surfaces.interactive_shell.runtime import answer_turn as cli_agent
 from surfaces.interactive_shell.runtime.answer_turn import answer_shell_question
 from surfaces.interactive_shell.session import Session
@@ -401,7 +401,7 @@ class TestAssistantOutputRendering:
 
         monkeypatch.setattr("core.llm.factory.get_llm", lambda _role: _Boom())
         monkeypatch.setattr(
-            "core.agent_harness.providers.default_providers.capture_exception",
+            "core.agent_harness.error_reporting.capture_exception",
             lambda exc, **_kwargs: captured_errors.append(exc),
         )
         session = Session()
